refactor(molecules): split molecule-display into focused modules
Extract card variants, shared helpers, and contributor empty-state UI so
molecule-display stays under 350 lines while preserving public exports.
