+++ src/components/molecules/molecule-display-constants.ts
@@ -1,0 +1,16 @@
+export const EDIT_FIELD_CLASS =
+  "rounded-lg border border-zinc-300 bg-zinc-100 px-3 py-2 text-sm focus:border-accent focus:ring-2 focus:ring-accent/20 dark:border-zinc-600 dark:bg-zinc-700/90";
+
+export const IDENTIFIER_CHIP_CLASS =
+  "rounded-md border border-zinc-200/80 bg-zinc-50/60 px-2 py-1.5 dark:border-zinc-600/60 dark:bg-zinc-700/30";
+
+export const CAS_FAVICON_URL =
+  "https://cdn.prod.website-files.com/650861f00f97fe8153979335/6585a20f2b9c762a8e082a87_cas-favicon.png";
+
+export const PUBCHEM_FAVICON_URL = "https://pubchem.ncbi.nlm.nih.gov/favicon.ico";
+
+export const COMPACT_MOLECULE_TAG_VISIBLE = 3;
+
+export const COPIED_RESET_MS = 2000;
+
+export const CAS_REGEX = /^\d{1,7}-\d{2}-\d$/;
